+++ Models/User.cs
@@ -6,4 +6,10 @@ public class User : IdentityUser
 {
     public ApiKey? ApiKey { get; set; } 
 
+    public List<Project> Projects { get; } = [];
+    
+    public ICollection<Asset> Assets { get; } = new List<Asset>(); // Collection navigation containing dependents
+
+    public ICollection<Comment> Comments { get; } = new List<Comment>(); // Collection navigation containing dependents
+
 }