+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project follows [Semantic Versioning](https://semver.org/).
 ### Added
 
 - Initialize new repo and add user authentication and authorization.
-  [#1](https://github.com/Medieinstitutet/ykweb25v1-the-assignment-thejoltjoker/pull/1)
+  [#1](https://github.com/Medieinstitutet/review-backend/pull/1)
+- Add models for assets, projects, comments and users.
+  [#2](https://github.com/Medieinstitutet/review-backend/pull/2)
 
 ### Changed